Don't swallow the underlying decrypt error
If key decryption fails, an error like:

```
Could not import SSH Key (Is it an RSA key? Is it password protected?)
```

is produced. It doesn't contain any info about what actually went wrong in decryption; the exception is swallowed.

Instead, report the real exception afterwards:

```
Could not import SSH Key (Is it an RSA key? Is it password protected?): Unsupport PEM encryption algorithm.
```